+++ crates/web-client/src/export.rs
@@ -92,10 +92,12 @@ impl WebClient {
                         ),
                     )
                 })?
-                .ok_or(JsValue::from_str("No account found"))?;
-            let account = account
-                .try_into()
-                .map_err(|_| JsValue::from_str("partial accounts are still unsupported"))?;
+                .ok_or_else(|| {
+                    JsValue::from_str(&format!(
+                        "Account with ID {} not found",
+                        account_id.to_string()
+                    ))
+                })?;
 
             let mut key_pairs = vec![];
 
@@ -115,7 +117,7 @@ impl WebClient {
             for commitment in commitments {
                 key_pairs.push(
                     keystore
-                        .get_key(commitment)
+                        .get_secret_key(commitment)
                         .await
                         .map_err(|err| {
                             js_error_with_context(err, "failed to get public key for account")